+++ TestMod/TestMod.cs
@@ -47,6 +47,7 @@ namespace TestMod
         private Skills.SkillType TestSkill;
 
         private ConfigEntry<KeyCode> EvilSwordSpecialConfig;
+        private ConfigEntry<InputManager.GamepadButton> EvilSwordGamepadConfig;
         private ButtonConfig EvilSwordSpecialButton;
         private CustomStatusEffect EvilSwordEffect;
 
@@ -96,6 +97,9 @@ namespace TestMod
 
             // Create a custom item with variants
             PrefabManager.OnVanillaPrefabsAvailable += AddCustomVariants;
+
+            // Crate a custom item with rendered icons
+            PrefabManager.OnVanillaPrefabsAvailable += AddItemsWithRenderedIcons;
 
             ZoneManager.OnVanillaLocationsAvailable += AddCustomLocations;
 
@@ -162,9 +166,10 @@ namespace TestMod
                 }
 
                 // Use the name of the ButtonConfig to identify the button pressed
-                if (EvilSwordSpecialButton != null && MessageHud.instance != null)
+                if (EvilSwordSpecialButton != null && MessageHud.instance != null && 
+                    Player.m_localPlayer != null && Player.m_localPlayer.m_visEquipment.m_rightItem == "EvilSword")
                 {
-                    if (ZInput.GetButtonDown(EvilSwordSpecialButton.Name) && MessageHud.instance.m_msgQeue.Count == 0)
+                    if (ZInput.GetButton(EvilSwordSpecialButton.Name) && MessageHud.instance.m_msgQeue.Count == 0)
                     {
                         MessageHud.instance.ShowMessage(MessageHud.MessageType.Center, "$evilsword_beevilmessage");
                     }
@@ -545,7 +550,9 @@ namespace TestMod
             // Add a client side custom input key for the EvilSword
             EvilSwordSpecialConfig = Config.Bind(JotunnTestModConfigSection, "EvilSwordSpecialAttack", KeyCode.B,
                 new ConfigDescription("Key to unleash evil with the Evil Sword"));
-
+            EvilSwordGamepadConfig = Config.Bind(JotunnTestModConfigSection, "EvilSwordSpecialAttackGamepad", InputManager.GamepadButton.ButtonSouth,
+                new ConfigDescription("Button to unleash evil with the Evil Sword"));
+            
         }
 
         // React on changed settings
@@ -597,7 +604,9 @@ namespace TestMod
             {
                 Name = "EvilSwordSpecialAttack",
                 Config = EvilSwordSpecialConfig,
-                HintToken = "$evilsword_beevil"
+                GamepadConfig = EvilSwordGamepadConfig,
+                HintToken = "$evilsword_beevil",
+                BlockOtherInputs = true
             };
             InputManager.Instance.AddButton(ModGUID, EvilSwordSpecialButton);
 
@@ -636,6 +645,12 @@ namespace TestMod
             Localization.AddTranslation("English", new Dictionary<string, string>
             {
                 { "lulz_shield", "Lulz Shield" }, { "lulz_shield_desc", "Lough at your enemies" }
+            });
+            
+            // Add translations for the rendered tree
+            Localization.AddTranslation("English", new Dictionary<string, string>
+            {
+                {"rendered_tree", "Rendered Tree"}, {"rendered_tree_desc", "A powerful tree, that can render its own icon. Magic!"}
             });
         }
 
@@ -848,7 +863,7 @@ namespace TestMod
             {
                 Item = "BlueprintTestRune"
             };
-            GUIManager.Instance.AddKeyHint(KHC_base);
+            KeyHintManager.Instance.AddKeyHint(KHC_base);
 
             // Add custom KeyHints for specific pieces
             KeyHintConfig KHC_make = new KeyHintConfig
@@ -861,7 +876,7 @@ namespace TestMod
                     new ButtonConfig { Name = "Attack", HintToken = "$bprune_make" }
                 }
             };
-            GUIManager.Instance.AddKeyHint(KHC_make);
+            KeyHintManager.Instance.AddKeyHint(KHC_make);
 
             KeyHintConfig KHC_piece = new KeyHintConfig
             {
@@ -873,7 +888,7 @@ namespace TestMod
                     new ButtonConfig { Name = "Attack", HintToken = "$bprune_piece" }
                 }
             };
-            GUIManager.Instance.AddKeyHint(KHC_piece);
+            KeyHintManager.Instance.AddKeyHint(KHC_piece);
 
             // Add additional localization manually
             Localization.AddTranslation("English", new Dictionary<string, string>
@@ -1166,7 +1181,7 @@ namespace TestMod
                         new ButtonConfig { Name = "Scroll", Axis = "Mouse ScrollWheel", HintToken = "$evilsword_scroll" }
                     }
                 };
-                GUIManager.Instance.AddKeyHint(KHC);
+                KeyHintManager.Instance.AddKeyHint(KHC);
             }
             catch (Exception ex)
             {
@@ -1265,6 +1280,43 @@ namespace TestMod
             {
                 // You want that to run only once, Jotunn has the item cached for the game session
                 PrefabManager.OnVanillaPrefabsAvailable -= AddCustomVariants;
+            }
+        }
+
+        // Create rendered icons from prefabs
+        private void AddItemsWithRenderedIcons()
+        {
+            try
+            {
+                // local function that will get called when the rendering is done
+                void CreateTreeItem(Sprite sprite)
+                {
+                    CustomItem treeItem = new CustomItem("item_MyTree", "BeechSeeds",
+                        new ItemConfig
+                        {
+                            Name = "$rendered_tree",
+                            Description = "$rendered_tree_desc",
+                            Icons = new[] { sprite },
+                            Requirements = new[]
+                            {
+                                new RequirementConfig { Item = "Wood", Amount = 1, Recover = true }
+                            }
+                        });
+                    ItemManager.Instance.AddItem(treeItem);
+                }
+
+                // use the vanilla beech tree prefab to render our icon from
+                GameObject beech = PrefabManager.Instance.GetPrefab("Beech1");
+                RenderManager.Instance.EnqueueRender(beech, CreateTreeItem);
+            }
+            catch (Exception ex)
+            {
+                Jotunn.Logger.LogError($"Error while adding item with rendering: {ex}");
+            }
+            finally
+            {
+                // You want that to run only once, Jotunn has the item cached for the game session
+                PrefabManager.OnVanillaPrefabsAvailable -= AddItemsWithRenderedIcons;
             }
         }